Linked all activities to main/ added agent functionality
changed the manifest to include the relations between activities and
added activities to display the list of agents/customers/packages/etc
including the layouts.

If you worked with a different name for the list page I can change the
name on the Manifest. I just wasn't sure what were you going to use.